requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
astroid==1.4.9
async-timeout==1.2.0
asyncio==3.4.3
beautifulsoup4==4.5.3
bleach==2.0.0
chardet==2.3.0
decorator==4.0.11
discord.py==0.16.7
entrypoints==0.2.2
html5lib==0.999999999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
multidict==2.1.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==4.4.1
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pylint==1.6.5
pyparsing==2.2.0
python-dateutil==2.6.0
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5
websockets==3.3
widgetsnbextension==2.0.0
wrapt==1.10.10
